3/2/EYAZIS/lab3/requirements.txt
spacy
nltk
lemminflect
pdfplumber
python-docx
striprtf
SQLAlchemy
pandas
PyQt6
tqdm
rich
wordfreq
fastapi
uvicorn
streamlit
requests
python-multipart
beautifulsoup4
psycopg2-binary
python-dotenv
